requirements.txt
# Dependências principais do projeto VaciVida AI
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50
openai>=1.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Azure Storage
azure-storage-blob>=12.19.0

# Dependências adicionais para análise de dados
pandas>=2.0.0
numpy>=1.24.0

# Para melhor performance
pydantic>=2.0.0
typing-extensions>=4.0.0

# Para deployment
gunicorn>=21.0.0
uvicorn>=0.24.0
